Estrutura criada. Já é possível cadastrar usuários com a senha já encriptada.
