+++ src/database/index.js
@@ -1,6 +1,11 @@
 const mongoose = require('mongoose');
 
-mongoose.connect('mongodb://localhost/noderest', { useMongoClient: true });
+// mongoose.set('useUnifiedTopology', true);
+mongoose.connect('mongodb://localhost:27017/noderest', 
+    { useNewUrlParser: true,
+    useUnifiedTopology: true,
+    useCreateIndex: true});
+
 mongoose.Promise = global.Promise;
 
 module.exports = mongoose;
