--- src/modules/user.js
@@ -1,30 +1,0 @@
-const mongoose = require('mongoose');
-
-const UserSchema = new mongoose.Schema({
-    name: {
-        type: String,
-        require: true
-    },
-    email: {
-        type: String,
-        unique: true,
-        required: true,
-        lowercase: true
-    },
-    password: {
-        type: String,
-        required: true,
-        select: false
-    },
-    birthDate: {
-        type: Date
-    },
-    createAt: {
-        type: Date,
-        default: Date.now
-    }
-});
-
-const User = mongoose.model('User', UserSchema);
-
-module.exports = User;
